fix: oversized icon images in search screen
Making the image resizable in ItemIconImageView made the icons in the search
screen look too big; the icons were occupying the whole space and as a consequence,
they lost their relative sizing.

An item with a really small icon (Air orb) was added to the mocks to better spot this.

TODO: implement a resizing mechanism that will preserve the relative sizing of items.